refactor: extract RocketChatUserClient from RocketChatService
